rework tests
